Rename `cupy.binary` submodule to `cupy._binary`
