--- cupy/binary/__init__.py
@@ -1,6 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.bitwise.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.binary import elementwise  # NOQA
-from cupy.binary import packing  # NOQA
